feat: Filter and set specific staff based on role in XrayStaff, PharmacyStaff, AttendentStaff components
This commit updates the XrayStaff, PharmacyStaff, and AttendentStaff components to filter the data received from the server based on the staff role. The code changes filter the data to only include staff members with the specified role (radiologist, pharmacist, and attendant respectively) and set the filtered data in the state variables. This enhancement improves the accuracy and relevance of the displayed staff data in the respective components.